# api-docs.json
{"openapi":"3.0.1","info":{"title":"T-RESERVE ENGINE API","description":"API для бронирования билетов","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Venues","description":"Площадки"},{"name":"Auth","description":"Регистрация, авторизация, refresh"},{"name":"Booking","description":"Ядро — бронирование мест"},{"name":"User Profile","description":"Профиль и бронирования пользователя"},{"name":"Events","description":"Мероприятия"},{"name":"Admin","description":"Управление мероприятиями (только ADMIN)"}],"paths":{"/api/admin/events/{id}":{"put":{"tags":["Admin"],"summary":"Редактировать мероприятие (до начала продаж)","operationId":"updateEvent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventUpdateRequest"}}},"required":true},"responses":{"400":{"description":"Невалидные данные или мероприятие уже началось","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}},"404":{"description":"Мероприятие не найдено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}},"403":{"description":"Доступ запрещён (требуется роль ADMIN)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}},"401":{"description":"Неавторизован (требуется JWT токен)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}},"200":{"description":"Мероприятие обновлено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}},"delete":{"tags":["Admin"],"summary":"Удалить мероприятие (каскадно билеты)","operationId":"deleteEvent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Мероприятие не найдено"},"403":{"description":"Доступ запрещён (требуется роль ADMIN)"},"409":{"description":"Нельзя удалить мероприятие с оплаченными билетами"},"401":{"description":"Неавторизован (требуется JWT токен)"},"204":{"description":"Мероприятие удалено"}}}},"/api/venues":{"get":{"tags":["Venues"],"summary":"Список площадок (ADMIN)","operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Venue"}}}}}}},"post":{"tags":["Venues"],"summary":"Создать площадку (ADMIN)","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Venue"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Venue"}}}}}}},"/api/bookings/{ticketId}/confirm":{"post":{"tags":["Booking"],"summary":"Подтвердить бронирование (mock оплата)","operationId":"confirm","parameters":[{"name":"ticketId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Билет не найден","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Неавторизован","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Попытка подтвердить чужое бронирование","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Бронирование подтверждено (LOCKED → BOOKED)","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Лок истёк или билет не в статусе LOCKED","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/bookings/lock":{"post":{"tags":["Booking"],"summary":"Заблокировать место (PG SELECT FOR UPDATE NOWAIT) → 200 OK / 409 Conflict","operationId":"lock","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LockRequest"}}},"required":true},"responses":{"409":{"description":"Место уже занято (race condition — другой пользователь успел первым)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LockResponse"}}}},"401":{"description":"Неавторизован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LockResponse"}}}},"200":{"description":"Место заблокировано, возвращает lockId и время истечения","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LockResponse"}}}},"400":{"description":"Невалидный запрос","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LockResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["Auth"],"summary":"Регистрация → access + refresh токены","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Пользователь зарегистрирован","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"400":{"description":"Невалидные данные (email, пароль)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"409":{"description":"Пользователь с таким email уже существует","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/refresh":{"post":{"tags":["Auth"],"summary":"Обновить access токен (передать refresh token)","operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"401":{"description":"Refresh token истёк или неверен","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"400":{"description":"Невалидный refresh token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"200":{"description":"Новая пара токенов","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["Auth"],"summary":"Логин → access + refresh токены","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Успешный вход","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"400":{"description":"Невалидные данные","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"401":{"description":"Неверный email или пароль","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/admin/events":{"post":{"tags":["Admin"],"summary":"Создать мероприятие (генерация билетов)","operationId":"createEvent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventCreateRequest"}}},"required":true},"responses":{"403":{"description":"Доступ запрещён (требуется роль ADMIN)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}},"401":{"description":"Неавторизован (требуется JWT токен)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}},"201":{"description":"Мероприятие создано","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}},"404":{"description":"Площадка не найдена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}},"400":{"description":"Невалидные данные запроса","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventResponse"}}}}}}},"/api/users/me":{"get":{"tags":["User Profile"],"summary":"Получить профиль текущего пользователя","operationId":"getProfile","responses":{"404":{"description":"Пользователь не найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"200":{"description":"Профиль получен","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}},"401":{"description":"Неавторизован (требуется JWT токен)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}}}}},"/api/users/me/bookings":{"get":{"tags":["User Profile"],"summary":"История бронирований пользователя (LOCKED + BOOKED)","operationId":"getMyBookings","responses":{"200":{"description":"Список бронирований получен","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserBookingResponse"}}}}},"401":{"description":"Неавторизован (требуется JWT токен)","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserBookingResponse"}}}}}}}},"/api/events":{"get":{"tags":["Events"],"summary":"Список мероприятий (пагинация)","operationId":"list_1","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"400":{"description":"Неверные параметры пагинации","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageEvent"}}}},"200":{"description":"Список мероприятий получен","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageEvent"}}}}}}},"/api/events/{id}":{"get":{"tags":["Events"],"summary":"Детали мероприятия","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Мероприятие не найдено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Event"}}}},"200":{"description":"Мероприятие найдено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Event"}}}}}}},"/api/events/{eventId}/seats":{"get":{"tags":["Events","Booking"],"summary":"Карта мест: все места + их статусы (polling каждые 3 сек, Redis cached)","operationId":"getSeats","parameters":[{"name":"eventId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Список мест с текущими статусами","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeatInfo"}}}}},"404":{"description":"Мероприятие не найдено","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeatInfo"}}}}}}}},"/api/admin/dashboard":{"get":{"tags":["Admin"],"summary":"Дашборд: статистика (mock)","operationId":"getDashboard","responses":{"403":{"description":"Доступ запрещён (требуется роль ADMIN)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardResponse"}}}},"401":{"description":"Неавторизован (требуется JWT токен)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardResponse"}}}},"200":{"description":"Статистика получена","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardResponse"}}}}}}},"/api/bookings/{ticketId}":{"delete":{"tags":["Booking"],"summary":"Ручная отмена блокировки","operationId":"cancel","parameters":[{"name":"ticketId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Билет не в статусе LOCKED"},"404":{"description":"Билет не найден"},"401":{"description":"Неавторизован"},"403":{"description":"Попытка отменить чужую блокировку"},"204":{"description":"Блокировка снята (LOCKED → AVAILABLE)"}}}}},"components":{"schemas":{"EventUpdateRequest":{"required":["status","title"],"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"startTime":{"type":"string","format":"date-time"},"basePrice":{"type":"number"},"imageUrl":{"type":"string"},"ageRestriction":{"type":"string"},"category":{"type":"string"},"durationMinutes":{"type":"integer","format":"int32"}}},"EventResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"imageUrl":{"type":"string"},"ageRestriction":{"type":"string"},"category":{"type":"string"},"durationMinutes":{"type":"integer","format":"int32"},"startTime":{"type":"string","format":"date-time"},"basePrice":{"type":"number"},"status":{"type":"string"},"venueId":{"type":"integer","format":"int64"},"venueName":{"type":"string"}}},"Venue":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"address":{"type":"string"},"rowsCount":{"type":"integer","format":"int32"},"colsCount":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"capacity":{"type":"integer","format":"int32"}}},"LockRequest":{"required":["eventId","seatId"],"type":"object","properties":{"eventId":{"type":"integer","format":"int64"},"seatId":{"type":"integer","format":"int64"}}},"LockResponse":{"type":"object","properties":{"lockId":{"type":"integer","format":"int64"},"expiresAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":100,"minLength":6,"type":"string"},"name":{"maxLength":100,"minLength":0,"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserInfo"}}},"UserInfo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"name":{"type":"string"},"role":{"type":"string"}}},"RefreshRequest":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"EventCreateRequest":{"required":["basePrice","startTime","title","venueId"],"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"venueId":{"type":"integer","format":"int64"},"startTime":{"type":"string","format":"date-time"},"basePrice":{"type":"number"},"imageUrl":{"type":"string"},"ageRestriction":{"type":"string"},"category":{"type":"string"},"durationMinutes":{"type":"integer","format":"int32"}}},"UserProfileResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"name":{"type":"string"},"role":{"type":"string"}}},"UserBookingResponse":{"type":"object","properties":{"ticketId":{"type":"integer","format":"int64"},"eventTitle":{"type":"string"},"seatLabel":{"type":"string"},"status":{"type":"string"},"price":{"type":"number"},"bookedAt":{"type":"string","format":"date-time"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"Event":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"venue":{"$ref":"#/components/schemas/Venue"},"title":{"type":"string"},"description":{"type":"string"},"imageUrl":{"type":"string"},"ageRestriction":{"type":"string"},"category":{"type":"string"},"durationMinutes":{"type":"integer","format":"int32"},"startTime":{"type":"string","format":"date-time"},"basePrice":{"type":"number"},"status":{"type":"string"},"version":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"PageEvent":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Event"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"SeatInfo":{"type":"object","properties":{"seatId":{"type":"integer","format":"int64"},"seatLabel":{"type":"string"},"rowLabel":{"type":"string"},"seatNumber":{"type":"integer","format":"int32"},"status":{"type":"string"},"price":{"type":"number"}}},"DashboardResponse":{"type":"object","properties":{"totalEvents":{"type":"integer","format":"int64"},"totalBookings":{"type":"integer","format":"int64"},"revenue":{"type":"number","format":"double"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"Введите JWT токен: Bearer {ваш_токен}","scheme":"bearer","bearerFormat":"JWT"}}}}